Aula 3 - Autenticação - Semana Python e Django da TreinaWeb.

+++ app/views/tarefa_views.py
@@ -1,17 +1,21 @@
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
 
 # Create your views here.
-from .entidades.tarefa import Tarefa
-from .form import TarefaForm
-from .services import tarefa_service
+from ..entidades.tarefa import Tarefa
+from ..form import TarefaForm
+from ..services import tarefa_service
 
 
+@login_required()
 def listar_tarefas(request):
-    tarefas = tarefa_service.listar_tarefas()
+    tarefas = tarefa_service.listar_tarefas(request.user)
     nome_tarefa = 'Assistir a Semana Python e Django da TreinaWeb'
     return render(request, 'tarefas/listar_tarefas.html', {"tarefas": tarefas})
 
 
+@login_required()
 def cadastrar_tarefa(request):
     if request.method == 'POST':
         form_tarefa = TarefaForm(request.POST)
@@ -23,7 +27,8 @@ def cadastrar_tarefa(request):
             tarefa_nova = Tarefa(titulo=titulo,
                                  descricao=descricao,
                                  data_expiracao=data_expiracao,
-                                 prioridade=prioridade)
+                                 prioridade=prioridade,
+                                 usuario=request.user)
             tarefa_service.cadastrar_tarefa(tarefa_nova)
             return redirect('listar_tarefas')
     else:
@@ -31,8 +36,11 @@ def cadastrar_tarefa(request):
     return render(request, 'tarefas/form_tarefa.html', {'form_tarefa': form_tarefa})
 
 
+@login_required()
 def editar_tarefa(request, id):
     tarefa_db = tarefa_service.listar_tarefa_id(id)
+    if tarefa_db.usuario != request.user:
+        return HttpResponse('Não permitido!')
     form_tarefa = TarefaForm(request.POST or None, instance=tarefa_db)
     if form_tarefa.is_valid():
         titulo = form_tarefa.cleaned_data['titulo']
@@ -42,14 +50,18 @@ def editar_tarefa(request, id):
         tarefa_nova = Tarefa(titulo=titulo,
                              descricao=descricao,
                              data_expiracao=data_expiracao,
-                             prioridade=prioridade)
+                             prioridade=prioridade,
+                             usuario=request.user)
         tarefa_service.editar_tarefa(tarefa_db, tarefa_nova)
         return redirect('listar_tarefas')
     return render(request, 'tarefas/form_tarefa.html', {'form_tarefa': form_tarefa})
 
 
+@login_required()
 def remover_tarefa(request, id):
     tarefa_db = tarefa_service.listar_tarefa_id(id)
+    if tarefa_db.usuario != request.user:
+        return HttpResponse('Não permitido!')
     if request.method == 'POST':
         tarefa_service.remover_tarefa(tarefa_db)
         return redirect('listar_tarefas')
